% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs-lwc.R
\name{A_func}
\alias{A_func}
\title{Step-2 accessibility \eqn{A} (per 10k)}
\usage{
A_func(data)
}
\arguments{
\item{data}{Table containing \code{R}, \code{GEOID}, and \code{p} (e.g. output of \code{\link[=R_func]{R_func()}}).}
}
\description{
Step-2 accessibility \eqn{A} (per 10k)
}
