% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs-lwc.R
\name{acs_gen}
\alias{acs_gen}
\title{Draw one truncated-normal ACS realization of drive / commute counts}
\usage{
acs_gen(data)
}
\arguments{
\item{data}{Output of \code{\link[=join_iso]{join_iso()}} (wide columns required).}
}
\description{
Draw one truncated-normal ACS realization of drive / commute counts
}
