% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs-lwc.R
\name{lwc_access}
\alias{lwc_access}
\title{Linearly weighted combination access: Monte Carlo over ACS MOE (\code{lwc_access})}
\usage{
lwc_access(data, pnts_c, pnts_w, iso_c, iso_w, n_sims, keep)
}
\arguments{
\item{data}{Tract attributes (see \code{\link[=join_iso]{join_iso()}}).}

\item{pnts_c, pnts_w}{Car / walk pharmacy \code{sf} with \code{query} and \code{within_isochrone}.}

\item{iso_c, iso_w}{Matching isochrone \code{sf} layers (\code{GEOID} column).}

\item{n_sims}{Number of simulation draws.}

\item{keep}{Character vector of \code{GEOID} values to retain in summaries.}
}
\description{
Linearly weighted combination access: Monte Carlo over ACS MOE (\code{lwc_access})
}
